+++ colossalai_platform/cli/api/__init__.py
@@ -1,14 +1,13 @@
-from .api import ColossalPlatformApi, StorageType, ApiError, LoginRequiredError, UploadRequest, DatasetNotFoundError, \
-    DatasetDeleteFilesRequest, NoObjectToDeleteError, DatasetListResponse
+from .api import ColossalPlatformApi
+from .dataset import DatasetListResponse, DatasetNotFoundError, DatasetDeleteFilesRequest, NoObjectToDeleteError
+from .types import ApiError, LoginRequiredError
 
 __all__ = [
     ColossalPlatformApi,
-    StorageType,
     ApiError,
     LoginRequiredError,
     DatasetNotFoundError,
     NoObjectToDeleteError,
-    UploadRequest,
     DatasetDeleteFilesRequest,
     DatasetListResponse,
 ]
