+++ colossalai_platform/cli/commands/dataset/upload_dir.py
@@ -4,7 +4,7 @@ import click
 import pathlib
 from typing import Iterable
 
-from colossalai_platform.cli.api import StorageType, UploadRequest, DatasetNotFoundError, DatasetDeleteFilesRequest, \
+from colossalai_platform.cli.api import DatasetNotFoundError, DatasetDeleteFilesRequest, \
     NoObjectToDeleteError
 from colossalai_platform.cli.context import CommandContext
 
@@ -27,7 +27,7 @@ def upload_dir(
 
     # check if the dataset exists
     try:
-        info = cmd_ctx.api.dataset_info(dataset_id)
+        info = cmd_ctx.api.dataset().info(dataset_id)
     except DatasetNotFoundError as e:
         return click.echo(e)
 
@@ -45,14 +45,14 @@ The dataset content would be overwritten.
 
     if not yes:
         y = click.prompt("Do you want to continue [y/N]")
-        if y != "y":
+        if y.lower() != "y":
             click.echo("Aborted!")
             return
 
     click.echo(f"Clearing dataset {dataset_id}...")
 
     try:
-        cmd_ctx.api.dataset_delete_files(DatasetDeleteFilesRequest(
+        cmd_ctx.api.dataset().delete_files(DatasetDeleteFilesRequest(
             datasetId=dataset_id,
             folders=[""],
         ))
@@ -66,12 +66,9 @@ The dataset content would be overwritten.
 
         click.echo(f"{local_file_path} => {storage_path}")
 
-        cmd_ctx.api.upload(
-            req=UploadRequest(
-                storage_type=StorageType.DATASET,
-                storage_id=dataset_id,
-                storage_path=storage_path,
-            ),
+        cmd_ctx.api.dataset().upload_local_file(
+            dataset_id=dataset_id,
+            storage_path=storage_path,
             local_file_path=local_file_path,
         )
